fix(studio): 完成系统审查并修复状态、安全与展示边界
