+++ src/studio/presentation/skill-detail-renderer.ts
@@ -1,48 +1,7 @@
+import { assessHealth, observedToolFailureRate } from '../application/skill-health.js';
 import { DEFAULT_LANG, e, layout } from './layout.js';
 import type { Lang } from '../../shared/language.js';
 import type { Insight, SkillIndexEntry } from '../view-models/index.js';
-
-export type HealthGrade = 'excellent' | 'good' | 'fair' | 'unhealthy' | 'unscored';
-
-export interface HealthAssessment {
-  grade: HealthGrade;
-  score: number | null;
-  label: string;
-  emoji: string;
-  color: 'green' | 'yellow' | 'red' | 'gray';
-}
-
-export function assessHealth(entry: SkillIndexEntry, insights: Insight[], lang: Lang): HealthAssessment {
-  const doctor = entry.doctor;
-  const doctorTotal = doctor === null ? 0 : doctor.passCount + doctor.warnCount + doctor.failCount;
-  const doctorScore = doctor !== null && doctorTotal > 0
-    ? ((doctor.passCount + doctor.warnCount * 0.5) / doctorTotal) * 100
-    : null;
-  const observeBand = entry.observe?.effectiveBand ?? 'gray';
-  const observeTrusted = observeBand !== 'gray';
-  const observeScore = observeTrusted ? (1 - entry.observe!.gapRate) * 100 : null;
-  const dimensions = [doctorScore, observeScore].filter((value): value is number => value !== null);
-  const score = dimensions.length === 0
-    ? null
-    : Math.round(dimensions.reduce((sum, value) => sum + value, 0) / dimensions.length);
-  const high = insights.some((insight) => insight.severity === 'high');
-  const medium = insights.some((insight) => insight.severity === 'medium');
-  const failed = (doctor?.failCount ?? 0) > 0 || observeBand === 'red';
-  const warned = (doctor?.warnCount ?? 0) > 0 || observeBand === 'yellow';
-  if (high || failed) {
-    return { grade: 'unhealthy', score, label: lang === 'zh' ? '不健康' : 'Unhealthy', emoji: '🔴', color: 'red' };
-  }
-  if (medium || warned) {
-    return { grade: 'fair', score, label: lang === 'zh' ? '待改进' : 'Fair', emoji: '🟡', color: 'yellow' };
-  }
-  if (score === null) {
-    return { grade: 'unscored', score: null, label: lang === 'zh' ? '未评估' : 'Unscored', emoji: '⚪', color: 'gray' };
-  }
-  if (insights.length === 0) {
-    return { grade: 'excellent', score, label: lang === 'zh' ? '健康' : 'Excellent', emoji: '🟢', color: 'green' };
-  }
-  return { grade: 'good', score, label: lang === 'zh' ? '良好' : 'Good', emoji: '🟢', color: 'green' };
-}
 
 function doctorSection(entry: SkillIndexEntry, lang: Lang): string {
   const doctor = entry.doctor;
@@ -54,10 +13,11 @@ function doctorSection(entry: SkillIndexEntry, lang: Lang): string {
 function observeSection(entry: SkillIndexEntry, lang: Lang): string {
   const observe = entry.observe;
   if (observe === null) return `<section class="sd-card"><h2>${lang === 'zh' ? '生产观察' : 'Observe'}</h2><p class="sd-muted">${lang === 'zh' ? '尚无生产观测。' : 'No production observations yet.'}</p></section>`;
+  const failureRate = observedToolFailureRate(observe);
   const confidence = observe.confidence === 'underpowered'
     ? (lang === 'zh' ? '样本不足，仅供参考' : 'Underpowered; indicative only')
     : observe.confidence;
-  return `<section class="sd-card"><h2>${lang === 'zh' ? '生产观察' : 'Observe'}</h2><dl class="sd-metrics"><div><dt>${lang === 'zh' ? '知识缺口' : 'Knowledge gap'}</dt><dd>${(observe.gapRate * 100).toFixed(1)}%</dd></div><div><dt>${lang === 'zh' ? '工具失败' : 'Tool failures'}</dt><dd>${(observe.failureRate * 100).toFixed(1)}%</dd></div><div><dt>${lang === 'zh' ? '片段数' : 'Segments'}</dt><dd>${observe.segmentCount}</dd></div><div><dt>${lang === 'zh' ? '可信度' : 'Confidence'}</dt><dd>${e(confidence)}</dd></div></dl></section>`;
+  return `<section class="sd-card"><h2>${lang === 'zh' ? '生产观察' : 'Observe'}</h2><dl class="sd-metrics"><div><dt>${lang === 'zh' ? '知识缺口' : 'Knowledge gap'}</dt><dd>${(observe.gapRate * 100).toFixed(1)}%</dd></div><div><dt>${lang === 'zh' ? '工具失败' : 'Tool failures'}</dt><dd>${failureRate === null ? (lang === 'zh' ? '未测得' : 'Not measured') : `${(failureRate * 100).toFixed(1)}%`}</dd></div><div><dt>${lang === 'zh' ? '片段数' : 'Segments'}</dt><dd>${observe.segmentCount}</dd></div><div><dt>${lang === 'zh' ? '可信度' : 'Confidence'}</dt><dd>${e(confidence)}</dd></div></dl></section>`;
 }
 
 function insightSection(insights: Insight[], lang: Lang): string {
