+++ test/studio/http/server-lifecycle.test.ts
@@ -2,13 +2,14 @@ import assert from 'node:assert/strict';
 import { mkdtempSync, rmSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
-import { afterEach, describe, it } from 'vitest';
+import { afterEach, describe, it, vi } from 'vitest';
 import { createReportServer } from '../../../src/studio/http/report-server.js';
 
 const temporaryDirectories: string[] = [];
 const runningServers: Array<ReturnType<typeof createReportServer>> = [];
 
 afterEach(async () => {
+  vi.unstubAllEnvs();
   await Promise.all(runningServers.splice(0).map((server) => server.stop()));
   for (const directory of temporaryDirectories.splice(0)) {
     rmSync(directory, { recursive: true, force: true });
@@ -28,6 +29,53 @@ async function waitUntilUnavailable(url: string): Promise<void> {
 }
 
 describe('Studio server lifecycle', () => {
+  it('serializes concurrent starts and a stop submitted while starting', async () => {
+    const observationsDir = mkdtempSync(join(tmpdir(), 'omk-studio-concurrent-'));
+    temporaryDirectories.push(observationsDir);
+    const server = createReportServer({ port: 0, observationsDir });
+    runningServers.push(server);
+    const first = server.start();
+    const second = server.start();
+    const stopped = server.stop();
+    const urls = await Promise.all([first, second]);
+    assert.equal(urls[0], urls[1]);
+    await stopped;
+    assert.equal(server.getUrl(), null);
+    await waitUntilUnavailable(urls[0]);
+    const restarted = await server.start();
+    assert.equal((await fetch(`${restarted}/health`)).status, 200);
+  });
+
+  it('does not expose catalog exceptions to the browser', async () => {
+    const observationsDir = mkdtempSync(join(tmpdir(), 'omk-studio-error-'));
+    temporaryDirectories.push(observationsDir);
+    const server = createReportServer({ port: 0, observationsDir, conversationCatalog: {
+      async listConversations() { throw new Error('token=secret /private/user/catalog'); },
+      async getConversation() { return undefined; },
+      async loadTaskTrajectory() { return undefined; },
+    } });
+    runningServers.push(server);
+    for (const path of ['/observe', '/api/conversations/activity']) {
+      const response = await fetch(`${await server.start()}${path}`);
+      assert.equal(response.status, 500);
+      assert.deepEqual(await response.json(), { error: 'studio_source_unavailable' });
+    }
+  });
+
+  it('can retry after a failed listen and formats a usable IPv6 URL', async () => {
+    const observationsDir = mkdtempSync(join(tmpdir(), 'omk-studio-listen-retry-'));
+    temporaryDirectories.push(observationsDir);
+    const server = createReportServer({ observationsDir, host: '::1' });
+    runningServers.push(server);
+    vi.stubEnv('OMK_REPORT_PORT', '-1');
+    await assert.rejects(server.start());
+    assert.equal(server.getUrl(), null);
+    vi.stubEnv('OMK_REPORT_PORT', '0');
+    const url = await server.start();
+    assert.match(url, /^http:\/\/\[::1\]:\d+$/);
+    assert.equal((await fetch(`${url}/health`)).status, 200);
+  });
+
   it('starts idempotently and shuts down after acknowledging the request', async () => {
     const observationsDir = mkdtempSync(join(tmpdir(), 'omk-studio-lifecycle-'));
     temporaryDirectories.push(observationsDir);
@@ -43,5 +91,8 @@ describe('Studio server lifecycle', () => {
     assert.equal(shutdown.status, 200);
     assert.deepEqual(await shutdown.json(), { ok: true });
     await waitUntilUnavailable(url);
+    assert.equal(server.getUrl(), null);
+    const restarted = await server.start();
+    assert.equal((await fetch(`${restarted}/health`)).status, 200);
   });
 });
