+++ test/studio/presentation/skill-health-confidence.test.ts
@@ -51,12 +51,12 @@ describe('skill-health renderer — confidence fallback for legacy reports', () 
   it('underpowered 顶部健康标签显示「样本不足」而非硬结论「需关注」', () => {
     const html = renderSkillHealthReport(legacyReport(3), 'zh');
     // 色带已静音为灰,标签也不能再报红色硬结论。
-    assert.match(html, /样本不足/);
+    assert.match(html, /class="rs-hero-grade"[^>]*>样本不足<\/div>/);
     assert.doesNotMatch(html, /需关注/);
   });
 
   it('high-N red overall 仍显示硬结论标签「需关注」', () => {
     const html = renderSkillHealthReport(legacyReport(40), 'zh');
-    assert.match(html, /需关注/);
+    assert.match(html, /class="rs-hero-grade"[^>]*>需关注<\/div>/);
   });
 });
